buzzer_shenanigans: actually fix pitches and support lower frequencies

+++ buzzer_shenanigans/src/buzzer.h
@@ -9,6 +9,6 @@ typedef struct buzzer_t {
 
 void buzzer_init(buzzer_t *bz, uint pin);
 void buzzer_deinit(buzzer_t *bz);
-void buzzer_play(buzzer_t *bz, uint32_t frequency, uint duration_ms);
+void buzzer_play(buzzer_t *bz, float frequency, uint duration_ms);
 
 #endif
